Face Emotion Detection, Pose Estimation and Red Dots on Green
I have added the codes for 3 project.
First: We can detect emotions of face of a person in the live video or any input MP4 video.
Second: Pose of person can be estimated in the live video or input video.
Third: This code detect every red dots in an input video and show green dots over them